fix code and add icon
added app icon and removed comments
